Project/django_test/requirements.txt
django==3.2.14
djangorestframework==3.13.1
Pillow==9.2.0
